feat: create RGBirdflop component to share with animtab
- Updated import paths for rgbStoreContext in multiple components to point to the new RGBirdflop component.
- Created a new RGBirdflop component to encapsulate the RGB functionality and state management.
- Removed the old rgb index file and integrated its functionality into the new RGBirdflop component.
- Ensured that all references to rgbStoreContext are consistent across the application.
