+++ src/components/Rgbirdflop/MobileNavbar.tsx
@@ -2,7 +2,7 @@ import { component$, useContext } from '@builder.io/qwik';
 import { Blend, Clipboard, FileJson, Palette, Settings, Sparkles } from 'lucide-icons-qwik';
 import { inlineTranslate } from 'qwik-speak';
 import { openItemsContext } from '~/routes/layout';
-import { rgbStoreContext } from '~/routes/resources/rgb';
+import { rgbStoreContext } from '~/components/Rgbirdflop/RGBirdflop';
 
 export default component$(({ animtab }: {
   animtab?: boolean;
