+++ src/components/Rgbirdflop/MyPrivatePresets.tsx
@@ -4,7 +4,7 @@ import { privatePresetsContext } from '~/routes/resources/rgb/presets';
 import PresetPreview from '~/components/Rgbirdflop/PresetPreview';
 import { CircleUserRound, Plus, Save, X } from 'lucide-icons-qwik';
 import { SelectMenu, Toggle } from '@luminescent/ui-qwik';
-import { renderPreview } from '~/routes/resources/rgb';
+import { renderPreview } from '~/components/Rgbirdflop/RGBirdflop';
 import { rgbDefaults } from '@birdflop/rgbirdflop';
 import { Form, Link } from '@builder.io/qwik-city';
 import { Notification, NotificationContext } from '~/util/Notification';
