+++ src/components/Rgbirdflop/TextShadow.tsx
@@ -1,5 +1,5 @@
 import { component$, useContext, useSignal } from '@builder.io/qwik';
-import { rgbStoreContext } from '~/routes/resources/rgb';
+import { rgbStoreContext } from '~/components/Rgbirdflop/RGBirdflop';
 import ColorMap from './ColorMap';
 import ColorList from './ColorList';
 import { inlineTranslate } from 'qwik-speak';
