+++ src/routes/resources/animpreview/index.tsx
@@ -4,7 +4,7 @@ import { getCookies, setCookies } from '~/util/dataUtils';
 import { inlineTranslate } from 'qwik-speak';
 import yaml from 'yaml';
 import Input, { previewStyleContext } from '~/components/Rgbirdflop/Input';
-import { rgbStoreContext } from '../rgb';
+import { rgbStoreContext } from '~/components/Rgbirdflop/RGBirdflop';
 import { Notification, NotificationContext } from '~/util/Notification';
 import { Eye } from 'lucide-icons-qwik';
 import { hexToRGB, rgbDefaults } from '@birdflop/rgbirdflop';
